chore: move utils/game_fight to core/game_fight, extract notifications/game_notifications from utils/game_utils
