+++ scripts/save2l/save2l.c
@@ -8,7 +8,6 @@ int G_DEBUG_ON = 0;
 int G_LOG_ON = 0;
 int G_EXPERIMENTAL_ON = 0;
 int G_FASTQUIT_ON = 0;
-int G_GODMODE_ON = 0;
 int G_DEBUG_ROOMTYPE = ENEMIES;
 int G_DEBUG_ROOMTYPE_ON = 0;
 char *G_DEBUG_ROOMTYPE_ARG = NULL;
