+++ src/utils/game_utils.c
@@ -2030,205 +2030,6 @@ void setConsumablePrices(int size, int *consumablePrices,
 }
 
 /**
- * Takes a Fighter pointer and an integer used to force execution.
- * If the force parameter is true, all checks are ignored.
- * If enemy's hp value is at least 50% of total, and none of atk, def or vel is 0 or less, nothing happens with an early return.
- * Otherwise, getBoost() is called to calc the level stat boost for each stat.
- * The BaseStats pointer for the fighter's fighterClass is loaded and each one of atk, def and vel is checked accounting for level boost.
- * If none of them is below the respective treshold of 35, 18 and 30 % of total, nothing happens.
- * Otherwise, all of them are reset to full amount accounting for permboosts and level boost.
- * @see Fighter
- * @see fighterClass
- * @see getBoost()
- * @param player The Fighter pointer to check the stats for.
- * @param force The integer to bypass all checks if true.
- */
-void statReset(Fighter *player, int force)
-{
-    log_tag("debug_log.txt", "[DEBUG]",
-            "Call to statReset() with ($force) == (%i)", force);
-    if (!force && (player->hp >= 0.5 * player->totalhp)
-        && !(player->atk <= 0 || player->def <= 0 || player->vel <= 0)) {
-        return;
-    }
-
-    int boost = getBoost(player->level, player->luck);
-
-    BaseStats *base = &basestats[player->class];
-    if (force || player->vel <= 0.3 * (base->vel + boost)
-        || player->atk <= 0.35 * (base->atk + boost)
-        || player->def <= 0.18 * (base->def + boost)) {
-        player->vel = base->vel + boost + player->permboost_vel;
-        player->atk = base->atk + boost + player->permboost_atk;
-        player->def = base->def + boost + player->permboost_def;
-        //Reset stats
-        if (!force) {
-            //yellow();
-            //printf("\n\n\t%s's stats reset.\n",player->name);
-            //white();
-        }
-    }
-}
-
-/**
- * Takes a Boss pointer and an integer used to force execution.
- * If the force parameter is true, all checks are ignored.
- * If boss's hp value is at least 40% of total, and none of atk, def or vel is 0 or less, nothing happens with an early return.
- * Otherwise, getBossBoost() is called to calc the level stat boost for each stat.
- * The BossBaseStats pointer for the boss's bossClass is loaded and each one of atk, def and vel is checked accounting for level boost.
- * If none of them is below the respective treshold of 30, 30 and 20 % of total, nothing happens.
- * Otherwise, all of them are reset to full amount accounting for beast boost and level boost.
- * @see Boss
- * @see bossClass
- * @see getBossBoost()
- * @see BSTFACTOR
- * @see BossBaseStats
- * @param b The Boss pointer to check the stats for.
- * @param force The integer to bypass all checks if true.
- */
-void statResetBoss(Boss *b, int force)
-{
-    if (!force && (b->hp >= 0.4 * b->totalhp)
-        && !(b->atk <= 0 || b->def <= 0 || b->vel <= 0)) {
-        return;
-    }
-    int boost = getBossBoost(b->level, b->class);
-
-    float beastf = 1;
-
-    if (b->beast) {
-        beastf = BSTFACTOR;
-    }
-    BossBaseStats *base = &basebossstats[b->class];
-
-    if (force) {		//We also update hp values
-        int hpBoost =
-            boost + round(base->level * 0.75) + (base->hp / 10) +
-            ((base->def / 4) % 10);
-        b->hp = round(beastf * (base->hp + hpBoost));
-        b->totalhp = b->hp;
-    }
-
-    if (force || b->vel <= (0.3 * (base->vel + boost))
-        || (b->atk <= (0.3 * (base->atk + boost)))
-        || b->def <= (0.2 * (base->def + boost))) {
-        b->vel = round(beastf * (base->vel + boost));
-        b->atk = round(beastf * (base->atk + boost));
-        b->def = round(beastf * (base->def + boost));
-        //Reset stats
-        if (!force) {
-            //yellow();
-            //printf("\n\n\t%s's stats reset.\n",stringFromEClass(e->class));
-            //white();
-        }
-    }
-}
-
-/**
- * Takes an Enemy pointer and an integer used to force execution.
- * If the force parameter is true, all checks are ignored.
- * If enemy's hp value is at least 40% of total, and none of atk, def or vel is 0 or less, nothing happens with an early return.
- * Otherwise, getEnemyBoost() is called to calc the level stat boost for each stat.
- * The EnemyBaseStats pointer for the enemy's enemyClass is loaded and each one of atk, def and vel is checked accounting for level boost.
- * If none of them is below the respective treshold of 30, 30 and 20 % of total, nothing happens.
- * Otherwise, all of them are reset to full amount accounting for beast boost and level boost.
- * @see Enemy
- * @see enemyClass
- * @see getEnemyBoost()
- * @see BSTFACTOR
- * @see EnemyBaseStats
- * @param e The Enemy pointer to check the stats for.
- * @param force The integer to bypass all checks if true.
- */
-void statResetEnemy(Enemy *e, int force)
-{
-    log_tag("debug_log.txt", "[DEBUG]",
-            "Call to statResetEnemy() with ($force) == (%i)", force);
-    if (!force && (e->hp >= 0.4 * e->totalhp)
-        && !(e->atk <= 0 || e->def <= 0 || e->vel <= 0)) {
-        return;
-    }
-    int boost = getEnemyBoost(e->level, e->class);
-
-    float beastf = 1;
-
-    if (e->beast) {
-        beastf = BSTFACTOR;
-    }
-    EnemyBaseStats *base = &baseenemystats[e->class];
-
-    if (force) {		//We also update hp values
-        int hpBoost =
-            floor(0.5 * boost + round(base->level * 0.75) + (base->hp / 10) +
-                  ((base->def / 4) % 10));
-        e->hp = round(beastf * (base->hp + hpBoost));
-        e->totalhp = e->hp;
-    }
-
-    if (force || e->vel <= (0.3 * (base->vel + boost))
-        || (e->atk <= (0.3 * (base->atk + boost)))
-        || e->def <= (0.2 * (base->def + boost))) {
-        e->vel = round(beastf * (base->vel + boost));
-        e->atk = round(beastf * (base->atk + boost));
-        e->def = round(beastf * (base->def + boost));
-        //Reset stats
-        if (!force) {
-            //yellow();
-            //printf("\n\n\t%s's stats reset.\n",stringFromEClass(e->class));
-            //white();
-        }
-    }
-}
-
-/**
- * Takes one integer and a bossClass and returns the boost relative to the level delta with base boss stats, as an integer.
- * The EnemyBossStats pointer for the boss's bossClass is loaded.
- * If the boost is negative, returns 0.
- * @see Boss
- * @see bossClass
- * @see BossBaseStats
- * @param lvl The level to check the boost against.
- * @param bclass The bossClass used to determine base level.
- * @return int The boost for any given stat, at the level passed as argument.
- */
-int getBossBoost(int lvl, bossClass bclass)
-{
-
-    BossBaseStats *base = &basebossstats[bclass];
-
-    float boost = ((lvl - base->level) * 1.25);
-    if (boost <= 0) {
-        boost = 0;
-    }
-
-    return (int)boost;
-}
-
-/**
- * Takes one integer and an enemyClass and returns the boost relative to the level delta with base enemy stats, as an integer.
- * The EnemyBaseStats pointer for the enemy's enemyClass is loaded.
- * If the boost is negative, returns 0.
- * @see Enemy
- * @see enemyClass
- * @see EnemyBaseStats
- * @param lvl The level to check the boost against.
- * @param eclass The enemyClass used to determine base level.
- * @return int The boost for any given stat, at the level passed as argument.
- */
-int getEnemyBoost(int lvl, enemyClass eclass)
-{
-
-    EnemyBaseStats *base = &baseenemystats[eclass];
-
-    float boost = ((lvl - base->level) * 1.25);
-    if (boost <= 0) {
-        boost = 0;
-    }
-
-    return (int)boost;
-}
-
-/**
  * Takes a Fighter pointer and resets all of its permboost_STAT values to 0, also correctly updating the current stat values.
  * @see Fighter
  * @param f The fighter pointer whose permboosts will be reset.
@@ -2611,284 +2412,6 @@ void printArtifactStats(Artifact *a)
     yellow();
     printf("\"%s\"\n", a->desc);
     white();
-}
-
-/**
- * Takes a RingaBuf pointer and queues the passed text to it.
- * @see handleRoom_Enemies()
- * @see handleRoom_Boss()
- * @param text The contents of the notification.
- * @param time The display time in milliseconds (unused, for now)
- * @param color The index of color to use for the notification drawing.
- * @param rb_notifications The Ringabuf to push the notification to.
- */
-void enqueue_notification(char *text, int time, int color, RingaBuf* rb_notifications)
-{
-    Notification notif = {0};
-
-    sprintf(notif.buf, "%s", text);
-    notif.color = color;
-
-#ifndef _WIN32
-    log_tag("debug_log.txt", "[DEBUG]", "%s():    pushing {%li} bytes to ringbuf", __func__, sizeof(notif));
-#else
-    log_tag("debug_log.txt", "[DEBUG]", "%s():    pushing {%lli} bytes to ringbuf", __func__, sizeof(notif));
-#endif
-    bool rb_res = rb_push(*rb_notifications, notif);
-
-
-    /* Log ring status buffer, from oldest to newest notification.
-    Notification* newest_notif = NULL;
-    Notification* oldest_notif = NULL;
-
-    if (!rb_notifications->is_full) {
-        log_tag("debug_log.txt", "[DEBUG]", "%s():    Logging up from 0 to head: { %" PRIu32 " }", __func__, rb_notifications->head);
-        for (int i = 0; i < (rb_notifications->head / sizeof(Notification)); i++) {
-            Notification* read_notif = (Notification*) &(rb_notifications->data[i * sizeof(Notification)]);
-            log_tag("debug_log.txt", "[DEBUG]", "%s():    Notification: [%s] Color: [%" PRId8 "]", __func__, read_notif->buf, read_notif->color);
-        }
-        oldest_notif = (Notification*) &(rb_notifications->data[0]);
-        newest_notif = (Notification*) &(rb_notifications->data[rb_notifications->head - (sizeof(Notification))]);
-    } else {
-        log_tag("debug_log.txt", "[DEBUG]", "%s():    Logging up from head+1 { %" PRIu32 " } to size { %" PRIu32 " }, then from 0 to head.", __func__, (rb_notifications->head / sizeof(Notification)) +1, rb_notifications->capacity / sizeof(Notification));
-        for (size_t i = (rb_notifications->head / sizeof(Notification)) +1; i < (rb_notifications->capacity / sizeof(Notification)); i++) {
-            Notification* read_notif = (Notification*) &(rb_notifications->data[i * sizeof(Notification)]);
-            log_tag("debug_log.txt", "[DEBUG]", "%s():    [%li] H+1->sz Notification: [%s] Color: [%" PRId8 "]", __func__, i, read_notif->buf, read_notif->color);
-        }
-        for (size_t i = 0; i < (rb_notifications->head / sizeof(Notification)); i++) {
-            Notification* read_notif = (Notification*) &(rb_notifications->data[i * sizeof(Notification)]);
-            log_tag("debug_log.txt", "[DEBUG]", "%s():    [%li] 0->H Notification: [%s] Color: [%" PRId8 "]", __func__, i, read_notif->buf, read_notif->color);
-        }
-
-        size_t newest_offset = (rb_notifications->head == 0 ? ((NOTIFICATIONS_RINGBUFFER_SIZE-1)* sizeof(Notification)) : (rb_notifications->head - sizeof(Notification)));
-        newest_notif = (Notification*) &(rb_notifications->data[newest_offset]);
-        oldest_notif = (Notification*) &(rb_notifications->data[(rb_notifications->head)]);
-
-    }
-
-    log_tag("debug_log.txt", "[DEBUG]", "%s():    Newest Notification: [%s] Color: [%" PRId8 "]", __func__, newest_notif->buf, newest_notif->color);
-    log_tag("debug_log.txt", "[DEBUG]", "%s():    Oldest Notification: [%s] Color: [%" PRId8 "]", __func__, oldest_notif->buf, oldest_notif->color);
-    */
-
-#ifndef _WIN32
-    if (!rb_res) log_tag("debug_log.txt", "[DEBUG]", "%s():    rb push failed. Head: { %" PRIu32 " } Tail: { %" PRIu32 " } Capacity: {%li}", __func__, rb_get_head(*rb_notifications), rb_get_tail(*rb_notifications), rb_get_capacity(*rb_notifications));
-#else
-    if (!rb_res) log_tag("debug_log.txt", "[DEBUG]", "%s():    rb push failed. Head: { %" PRIu32 " } Tail: { %" PRIu32 " } Size: {%lli}", __func__, rb_get_head(*rb_notifications), rb_get_tail(*rb_notifications), rb_get_capacity(*rb_notifications));
-#endif
-
-    /*
-    wprintw(w, "\n  %s", text);
-    wrefresh(w);
-    //refresh();
-    napms(time);
-    */
-}
-
-/**
- * Takes a RingaBuf pointer to queue notifications to, a Fighter pointer value and applies the effect pertaining to its status value.
- * @see Fighter
- * @see fighterStatus
- * @see printStatusText()
- * @param f The Fighter pointer at hand.
- * @param rb_notifications The RingaBuf pointer used for notifications.
- */
-void applyStatus(Fighter *f, RingaBuf* rb_notifications)
-{
-
-    switch (f->status) {
-    case Normal: {
-        break;
-    }
-    break;
-    case Poison: {
-
-        //Account for penicillin perk
-        //ATM multiples don't stack
-        int penicillin = f->perks[PENICILLIN]->innerValue;
-        if (penicillin > 0) {
-            return;
-        }
-
-        if (f->hp >= 4) {
-            f->hp -= 3;
-        } else {
-            f->hp = 1;	//Will this be a problem?
-        }
-        printStatusText(Poison, f->name, S4C_RED, rb_notifications);
-    }
-    break;
-    case Burned: {
-        printStatusText(Burned, f->name, S4C_RED, rb_notifications);
-    }
-    break;
-    case Frozen: {
-        printStatusText(Frozen, f->name, S4C_RED, rb_notifications);
-    }
-    break;
-    case Weak:
-
-        break;
-    case Strong:
-
-        break;
-    }
-}
-
-/**
- * Takes a RingaBuf pointer to queue notifications to, a Enemy pointer value and applies the effect pertaining to its status value.
- * @see Enemy
- * @see fighterStatus
- * @see printStatusText()
- * @see stringFromEClass()
- * @param e The Enemy pointer at hand.
- * @param rb_notifications The RingaBuf used for notifications.
- * @see enqueue_notification()
- */
-void applyEStatus(Enemy *e, RingaBuf* rb_notifications)
-{
-    switch (e->status) {
-    case Normal: {
-        break;
-    }
-    break;
-    case Poison: {
-        if (e->hp >= 4) {
-            e->hp -= 3;
-        } else {
-            e->hp = 1;	//Will this be a problem for kills in the enemy loop?
-        }
-        printStatusText(Poison, stringFromEClass(e->class), S4C_BRIGHT_GREEN, rb_notifications);
-    }
-    break;
-    case Burned: {
-        if (e->hp >= 5) {
-            e->hp -= 4;
-        } else {
-            e->hp = 1;	//Will this be a problem for kills in the enemy loop?
-        }
-
-        if (e->atk >= 3) {
-            e->atk -= 3;
-        } else {
-            e->atk = 1;
-        }
-        printStatusText(Burned, stringFromEClass(e->class), S4C_BRIGHT_GREEN, rb_notifications);
-    }
-    break;
-    case Frozen: {
-        if (e->vel >= 3) {
-            e->vel -= 1;
-        } else {
-            e->vel = 1;	//Will this be a problem for kills in the enemy loop?
-        }
-        printStatusText(Frozen, stringFromEClass(e->class), S4C_BRIGHT_GREEN, rb_notifications);
-    }
-
-    break;
-    case Weak:
-
-        break;
-    case Strong:
-
-        break;
-    }
-}
-
-/**
- * Takes a RingaBuf pointer to queue notifications to, a Boss pointer value and applies the effect pertaining to its status value.
- * @see Boss
- * @see fighterStatus
- * @see printStatusText()
- * @see stringFromBossClass()
- * @param b The Boss pointer at hand.
- * @param rb_notifications The RingaBuf used for notifications.
- */
-void applyBStatus(Boss *b, RingaBuf* rb_notifications)
-{
-    switch (b->status) {
-    case Normal: {
-        break;
-    }
-    break;
-    case Poison: {
-        if (b->hp >= 4) {
-            b->hp -= 3;
-        } else {
-            b->hp = 1;	//Will this be a problem for kills in the enemy loop?
-        }
-        printStatusText(Poison, stringFromBossClass(b->class), S4C_BRIGHT_GREEN, rb_notifications);
-    }
-    break;
-    case Burned: {
-        if (b->hp >= 5) {
-            b->hp -= 4;
-        } else {
-            b->hp = 1;	//Will this be a problem for kills in the enemy loop?
-        }
-
-        if (b->atk >= 3) {
-            b->atk -= 3;
-        } else {
-            b->atk = 1;
-        }
-        printStatusText(Burned, stringFromBossClass(b->class), S4C_BRIGHT_GREEN, rb_notifications);
-    }
-    break;
-    case Frozen: {
-        if (b->vel >= 3) {
-            b->vel -= 1;
-        } else {
-            b->vel = 1;	//Will this be a problem for kills in the enemy loop?
-        }
-        printStatusText(Frozen, stringFromBossClass(b->class), S4C_BRIGHT_GREEN, rb_notifications);
-    }
-
-    break;
-    case Weak:
-
-        break;
-    case Strong:
-
-        break;
-    }
-}
-
-/**
- * Takes a RingaBuf pointer to queue notifications to, a fighterStatus value and a string of who's the entity to print the respective status message.
- * @see fighterStatus
- * @param status The fighterStatus at hand.
- * @param subject A string with name of entity owning the fighterStatus.
- * @param color The color to use for the notification.
- * @param rb_notifications The RingaBuf used for notifications.
- */
-void printStatusText(fighterStatus status, char *subject, int color, RingaBuf* rb_notifications)
-{
-    char msg[500];
-    switch (status) {
-    case Normal: {
-        return;
-    };
-    break;
-    case Poison:
-    case Burned: {
-        sprintf(msg, "%s is hurt by its %s.", subject,
-                stringFromStatus(status));
-        enqueue_notification(msg, 500, color, rb_notifications);
-    }
-    break;
-    case Weak:
-    case Strong: {
-        sprintf(msg, "%s is feeling %s.", subject,
-                stringFromStatus(status));
-        enqueue_notification(msg, 500, color, rb_notifications);
-    }
-    break;
-    case Frozen: {
-        sprintf(msg, "%s is frozen cold.", subject);
-        enqueue_notification(msg, 500, color, rb_notifications);
-    }
-    break;
-    }
 }
 
 /**
@@ -4519,25 +4042,6 @@ int onLevelUp(Fighter *player)
 }
 
 /**
- * Takes two integers for level to calc against and luck, and returns the boost relative to the level with luck variations, as an integer.
- * At level 1, returns 0.
- * @param lvl The level to check the boost against.
- * @param luck The luck value to influence calcs.
- * @return int The boost for any given stat, at the level passed as argument.
- */
-int getBoost(int lvl, int luck)
-{
-
-    float boost = (lvl * 1.25F) + ((luck % 2) * 2);
-
-    if (lvl < 2) {
-        boost = 1.0;		// Evitare conflitti
-    }
-
-    return (int)boost;
-}
-
-/**
  * Takes a Fighter pointer and deleted all the equips not in use, granting a payment to the Fighter balance.
  * @see Shop
  * @see handleRoom_Shop()
@@ -4804,23 +4308,6 @@ void quit(Fighter *p, Room *room, loadInfo *load_info, Koliseo_Temp *t_kls)
     //freeRoom(room);
     log_tag("debug_log.txt", "[DEBUG]", "Quitting program.");
     exit(EXIT_SUCCESS);
-}
-
-/**
- * Takes a Turncounter pointer and an integer.
- * If the count value at the pointer is 0 (counter is inactive), the turns valueis assigned.
- * @see Turncounter
- * @param c The Turncounter whose count value will be set.
- * @param turns The value to be assigned.
- */
-void setCounter(Turncounter *c, int turns)
-{
-
-    if (c->count == 0) {	// Counter is NOT already active
-        c->count = turns;
-    } else {
-        //Handle counters already activ
-    }
 }
 
 /**
